moved signout button to navbar
